[s3proxy] move the `AllwaysAllow` setup to the test files
